source_code/src/requirements.txt
alembic==1.8.1    
cachetools==5.2.0    
certifi==2022.6.15
charset-normalizer==2.1.1    
click==8.1.3    
colorama==0.4.5    
Flask==2.2.2    
redis==4.3.4
Flask-Migrate==3.1.0    
Flask-SQLAlchemy==2.5.1
mysqlclient==2.0.3    
psycopg2-binary==2.8.6
google-api-core==2.10.0   
google-api-python-client==2.58.0   
google-auth==2.11.0   
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.3
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
msgpack-python==0.5.6
nanomsg==1.0
nanoservice==0.7.2
nose==1.3.7
numpy==1.23.2
oauthlib==3.2.0
oi==0.4.2
pandas==1.4.4
pip==22.2.2
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==63.4.3
six==1.16.0
SQLAlchemy==1.4.40
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
